Graph mode supports all Op test
