+++ python/oneflow/test/modules/test_diag.py
@@ -40,7 +40,7 @@ class Test_Diag_module(flow.unittest.TestCase):
         x = random_tensor(ndim=2, dim0=random(), dim1=random()).to(device)
         return torch.diag(x)
 
-    @autotest(auto_backward=False, check_graph=False)
+    @autotest(auto_backward=False, check_graph=True)
     def test_diag_one_dim(test_case):
         device = random_device()
         x = random_tensor(ndim=1, dim0=random()).to(device, torch.bool)
