+++ python/oneflow/test/modules/test_logical_and.py
@@ -86,7 +86,7 @@ class TestLogicalAndModule(flow.unittest.TestCase):
         y = torch.logical_and(x1, x2)
         return y
 
-    @autotest(n=10, auto_backward=False, check_graph=False)
+    @autotest(n=10, auto_backward=False, check_graph=True)
     def test_logical_and_bool_with_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
