+++ python/oneflow/test/modules/test_logical_not.py
@@ -59,7 +59,7 @@ class TestLogicalNotModule(flow.unittest.TestCase):
         y = torch.logical_not(x1)
         return y
 
-    @autotest(n=10, auto_backward=False, check_graph=False)
+    @autotest(n=10, auto_backward=False, check_graph=True)
     def test_logical_not_bool_with_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
