+++ python/oneflow/test/modules/test_logical_reduce.py
@@ -43,7 +43,7 @@ class TestLogicalReduce(flow.unittest.TestCase):
         )
         return torch.all(x, dim)
 
-    @autotest(auto_backward=False, check_graph=False)
+    @autotest(auto_backward=False, check_graph=True)
     def test_any_with_random_data(test_case):
         device = random_device()
         dim = random(1, 4).to(int)
@@ -59,7 +59,7 @@ class TestLogicalReduce(flow.unittest.TestCase):
         )
         return torch.any(x, dim)
 
-    @autotest(auto_backward=False, check_graph=False)
+    @autotest(auto_backward=False, check_graph=True)
     def test_scalar_reduce_all_with_random_data(test_case):
         device = random_device()
         x = random_tensor(ndim=4, dtype=float, requires_grad=False).to(device)
