+++ python/oneflow/test/modules/test_logical_xor.py
@@ -106,7 +106,7 @@ class TestLogicalXorModule(flow.unittest.TestCase):
         y = torch.logical_xor(x1, x2)
         return y
 
-    @autotest(n=10, auto_backward=False, check_graph=False)
+    @autotest(n=10, auto_backward=False, check_graph=True)
     def test_logical_xor_bool_with_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
