+++ python/oneflow/test/tensor/test_parameter.py
@@ -25,14 +25,14 @@ from oneflow.test_utils.automated_test_util import *
 
 @flow.unittest.skip_unless_1n1d()
 class TestParameter(flow.unittest.TestCase):
-    @autotest(n=1, check_graph=False)
+    @autotest(n=1, check_graph=True)
     def test_parameter_grad_fn_none(test_case):
         x = torch.ones(2, 3).requires_grad_(True)
         y = x + x
         z = torch.nn.Parameter(y)
         return z.grad_fn
 
-    @autotest(n=1, check_graph=False)
+    @autotest(n=1, check_graph=True)
     def test_parameter_set_data_autograd_meta(test_case):
         x = torch.ones(2, 3).requires_grad_(True)
         y = x + x
